refactor(activate): activation_diff.rs -> diff_serializer.rs
I think this was missed when the DiffSerializer struct was renamed

I also used an import instead of activation_diff::FLOX_HOOK_DIFF_VAR
