+++ cli/flox-activations/src/diff_serializer.rs
