+++ cli/flox-activations/src/lib.rs
@@ -1,7 +1,7 @@
-pub mod activation_diff;
 pub mod attach;
 pub mod attach_diff;
 pub mod cli;
+pub mod diff_serializer;
 pub mod env_diff;
 pub mod gen_rc;
 pub mod hook;
